Add files via upload
Programs I wrote for the first test in Programming for Informatics.

Question 1- Count the number of iterations (i.e. the number of times it generates a new random integer). Your program should update and print out the highest number (i.e. the maximum number) generated so far for each iteration. [For example, if the first random number is 89 and the next generated random number is 112, then "highest number generated" would be "112." If the following random number generated is 16, then "highest number generated" would still be "112."] When 44 is generated, the total number of iterations should be printed and the loop should terminate. In addition, run your program three times and report the results (the largest number generated and the total number of iterations) as comments in your file.

Question 2- Using a while loop, write a program that asks the user to enter a number in the range 205 through 259 and validates the input (i.e. continually asks the user for a number between 205 and 259 until the user puts in the correct input.) If invalid input is entered (just validate that input is within the range), print an error message and ask the user (again) to enter a number in the range 205 through 259.

Question 3- Ask the user for the number of words that they will enter. Using a for loop, write Python code that asks the user to enter a single word at a time. The loop should iterate exactly the number of times indicated by the user input (the number of words the user wants to enter), each time asking for a single word and adding all the words together by appending the new words with a space between each word entered (similar to writing a single sentence). Before ending the program, print the sentence out.

Question 4- A video download service offers discounts for customers who use the service more frequently. Individual video downloads are always $3.89. Quantity discounts are determined at the end of the user's annual term according to the following table:

    Total number of videos downloaded/year    |    Discount applied
    Less than 100                                                 |    No discount
    100 to 179                                                       |    4%
    180 to 339                                                      |    7.5%
    340 or more (not more than 600)                |    9%

Accept the total number of downloaded videos for the year from the user. If the number entered is outside the acceptable range, the program should display an error message and exit. Write a function that passes this value and returns the discount to be applied. Determine then display the customer's annual savings based on the discount they received.