Add files via upload
Programs I wrote for the second test in Programming for Informatics.

Question 1- Write a program to read through a file of a winning lottery numbers. Open the text file Lottery_Winning_Numbers.txt in your program; examine the text file to determine the format of the text on each line of the file; there are 3 fields of data per line. Create a data structure  (use a dictionary) with a single (2 digit) lottery number as the dictionary key and the number of times the number appears in this file (i.e. the number of times a single 2 digit lottery number appears in every "winning lottery number" listed in this file) as the dictionary value.

Output

-The program should print the total number of the "winning lottery numbers" processed. (There is one complete "winning lottery number" per line.)
- The program should then print (on separate lines) all the individual (unique) 2 digit numbers in the file and the number of times they appeared.
- Each series of "winning lottery numbers" looks like this: 09    36    44    53    59. So each of these lottery numbers are made up of five 2 digit numbers (e.g. "09", "36") between "01" and "60" (inclusive). These should be sorted from the smallest lottery number to the largest lottery number.

Use only the dictionary methods and functions covered in this course.
Add comments throughout your code describing what the program is doing.